refactor (api-cat) Rename api-cat-client to api-cat-bindings
