+++ applications/api-cat/pom.xml
@@ -134,7 +134,7 @@
 
         <dependency>
             <groupId>no.dcat</groupId>
-            <artifactId>api-cat-client</artifactId>
+            <artifactId>api-cat-bindings</artifactId>
             <version>1.1.0</version>
         </dependency>
         <dependency>
