+++ applications/api-cat/src/main/java/no/acat/controller/ConvertController.java
@@ -4,8 +4,8 @@ import com.google.common.base.Strings;
 import io.swagger.v3.oas.models.OpenAPI;
 import no.acat.service.ParserService;
 import no.acat.spec.ParseException;
-import no.dcat.client.apicat.ConvertRequest;
-import no.dcat.client.apicat.ConvertResponse;
+import no.fdk.acat.bindings.ConvertRequest;
+import no.fdk.acat.bindings.ConvertResponse;
 import no.dcat.webutils.exceptions.BadRequestException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
