+++ applications/api-cat/src/test/java/no/acat/controller/ConvertControllerTest.java
@@ -3,8 +3,8 @@ package no.acat.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import no.acat.service.ParserService;
 import no.acat.utils.Utils;
-import no.dcat.client.apicat.ConvertRequest;
-import no.dcat.client.apicat.ConvertResponse;
+import no.fdk.acat.bindings.ConvertRequest;
+import no.fdk.acat.bindings.ConvertResponse;
 import no.dcat.shared.testcategories.UnitTest;
 import no.dcat.webutils.exceptions.BadRequestException;
 import org.junit.Assert;
