+++ applications/registration-api/pom.xml
@@ -213,7 +213,7 @@
         </dependency>
         <dependency>
             <groupId>no.dcat</groupId>
-            <artifactId>api-cat-client</artifactId>
+            <artifactId>api-cat-bindings</artifactId>
             <version>1.1.0</version>
         </dependency>
         <dependency>
