+++ libraries/api-cat-bindings/src/main/java/no/fdk/acat/bindings/ConvertRequest.java
@@ -1,4 +1,4 @@
-package no.dcat.client.apicat;
+package no.fdk.acat.bindings;
 
 
 import lombok.Builder;
