+++ libraries/api-cat-bindings/src/main/java/no/fdk/acat/bindings/ConvertResponse.java
@@ -1,4 +1,4 @@
-package no.dcat.client.apicat;
+package no.fdk.acat.bindings;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import lombok.Data;
